docs(claude): page/component data access layer rules 명문화
