+++ web/src/actions/tags/constants.ts
@@ -1,4 +1,4 @@
 export {
 	PROBLEM_BY_TAG_SORT_KEYS,
 	type ProblemByTagSort,
-} from "@/lib/services/problem-vote-tags";
+} from "@/lib/services/problem-list-sort";
